requirements.txt
beautifulsoup4==4.12.2
pandas==2.0.3
Requests==2.31.0
SQLAlchemy==2.0.18
wget==3.2
python-dotenv==1.0.0
SQLAlchemy-Utils==0.41.1
lxml==4.9.3
nltk==3.8.1
django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6